+++ methane_sensor_v0/HAL/UART1/UART1_HAL.c
@@ -20,8 +20,8 @@
 static void copy_buffer(uint8_t msg[]);
 
 static uint8_t tx_buffer[UART1_TX_BUFF_LENGTH];
-static uint8_t tx_buffer_data_len=0;
-static uint8_t tx_buffer_cursor=0;
+static uint16_t tx_buffer_data_len=0;
+static uint16_t tx_buffer_cursor=0;
 
 static uint8_t rx_buffer[UART1_RX_BUFF_LENGTH];
 static uint8_t rx_buffer_data_len=0;
@@ -40,7 +40,7 @@ void uart1_hal_init(){
 void uart1_hal_send_string(const char msg[]){
 	while(tx_buffer_data_len!=0){};//wait to write if buffer not empty
 	uint16_t len=strlen(msg);
-	for (uint8_t i=0; i<len; i++)
+	for (uint16_t i=0; i<len; i++)
 	{
 		tx_buffer[i]=(uint8_t)msg[i];
 	}
@@ -54,7 +54,7 @@ void uart1_hal_send_string(const char msg[]){
 
 void uart1_hal_send_message(uint8_t msg[], uint8_t length){
 	while(tx_buffer_data_len!=0){};//wait to write if buffer not empty
-	for (uint8_t i=0; i<length; i++)
+	for (uint16_t i=0; i<length; i++)
 	{
 		tx_buffer[i]=msg[i];
 	}
